app/live2d/renderer.py
from OpenGL.GL import *
from OpenGL.GLU import *
import numpy as np
from PIL import Image
import os

class Live2DRenderer:
    """OpenGL renderer for Live2D models"""
    
    def __init__(self):
        self.textures = {}
        self.shader_program = None
        self.width = 0
        self.height = 0
    
    def init_gl(self):
        """Initialize OpenGL resources"""
        # Initialize shaders
        self.init_shaders()
    
    def init_shaders(self):
        """Initialize GLSL shaders for Live2D rendering"""
        # In a real implementation, this would compile and link shaders
        # For now, this is a placeholder
        
        # Vertex shader source
        vertex_shader_source = """
        #version 330 core
        layout (location = 0) in vec3 aPos;
        layout (location = 1) in vec2 aTexCoord;
        
        out vec2 TexCoord;
        
        uniform mat4 model;
        uniform mat4 projection;
        
        void main()
        {
            gl_Position = projection * model * vec4(aPos, 1.0);
            TexCoord = aTexCoord;
        }
        """
        
        # Fragment shader source
        fragment_shader_source = """
        #version 330 core
        out vec4 FragColor;
        
        in vec2 TexCoord;
        
        uniform sampler2D texture1;
        uniform vec4 color;
        
        void main()
        {
            FragColor = texture(texture1, TexCoord) * color;
        }
        """
        
        # In a real implementation, these would be compiled and linked
        # self.shader_program = compile_and_link_shaders(vertex_shader_source, fragment_shader_source)
    
    def load_texture(self, texture_path):
        """Load a texture from file"""
        if texture_path in self.textures:
            return self.textures[texture_path]
        
        try:
            # Load image using PIL
            image = Image.open(texture_path)
            image_data = np.array(list(image.getdata()), np.uint8)
            
            # Generate texture
            texture_id = glGenTextures(1)
            glBindTexture(GL_TEXTURE_2D, texture_id)
            
            # Set texture parameters
            glTexParameteri(GL_TEXTURE_2D, GL_TEXTURE_WRAP_S, GL_CLAMP_TO_EDGE)
            glTexParameteri(GL_TEXTURE_2D, GL_TEXTURE_WRAP_T, GL_CLAMP_TO_EDGE)
            glTexParameteri(GL_TEXTURE_2D, GL_TEXTURE_MIN_FILTER, GL_LINEAR)
            glTexParameteri(GL_TEXTURE_2D, GL_TEXTURE_MAG_FILTER, GL_LINEAR)
            
            # Upload texture data
            width, height = image.size
            if image.mode == "RGBA":
                glTexImage2D(GL_TEXTURE_2D, 0, GL_RGBA, width, height, 0, GL_RGBA, GL_UNSIGNED_BYTE, image_data)
            else:
                # Convert to RGBA if needed
                image = image.convert("RGBA")
                image_data = np.array(list(image.getdata()), np.uint8)
                glTexImage2D(GL_TEXTURE_2D, 0, GL_RGBA, width, height, 0, GL_RGBA, GL_UNSIGNED_BYTE, image_data)
            
            # Store and return texture ID
            self.textures[texture_path] = texture_id
            return texture_id
        except Exception as e:
            print(f"Failed to load texture {texture_path}: {e}")
            return 0
    
    def resize(self, width, height):
        """Handle viewport resize"""
        self.width = width
        self.height = height
        
        # Update projection matrix
        # In a real implementation, this would update the shader uniform
    
    def begin_render(self):
        """Begin rendering frame"""
        glClear(GL_COLOR_BUFFER_BIT | GL_DEPTH_BUFFER_BIT)
        
        # In a real implementation, this would activate shaders and set up matrices
        # glUseProgram(self.shader_program)
    
    def end_render(self):
        """End rendering frame"""
        # In a real implementation, this would clean up after rendering
        # glUseProgram(0)
    
    def render_model(self, model_data):
        """Render a Live2D model"""
        # This would use the Live2D SDK to render the model
        # For now, this is a placeholder
        pass
    
    def cleanup(self):
        """Clean up OpenGL resources"""
        # Delete textures
        for texture_id in self.textures.values():
            glDeleteTextures(1, [texture_id])
        self.textures.clear()
        
        # Delete shader program
        if self.shader_program:
            # glDeleteProgram(self.shader_program)
            self.shader_program = None 