requirements.txt
PySide6>=6.4.0
PyOpenGL>=3.1.6
PyOpenGL-accelerate>=3.1.6
numpy>=1.23.0
Pillow>=9.2.0
python-dotenv>=0.21.0
psutil>=5.9.0
requests>=2.28.0
pyobjc-framework-Cocoa>=8.5.0; sys_platform == 'darwin'  # For macOS transparency support
pywin32>=303; sys_platform == 'win32'  # For Windows specific features
